pkgs/cli: Hide `clan history` subcommand from help output
Hide the `clan history` subcommand from the help output.
It is only to be used by *very* advanced users currently and not even
them.
It is confusing in the current help output, because it is not clear what
it does.

This keeps the functionality around - while not confusing new users.